Add files via upload
Round Manager.cs
Allows you to start a new round, and completes the round when all of the enemies have been destroyed. Also tells each spawner how many enemies they will be spawning.

Spawner.cs
Spawns in an enemy prefab at the transforms position.

EnemyAi.cs
Used for testing that that RoundSpawner.cs was working. In your enemy script, make sure to add a reference to the RoundManger, and to call the RemoveEnemy() method located in the RoundManager.